+++ docs/pages/documentation/data/table/api.js
@@ -157,6 +157,11 @@ export default [
                 name: '<code>empty</code>',
                 description: 'Replaces table body when <code>data</code> array prop is empty',
                 props: ''
+            },
+            {
+                name: '<code>footer</code>',
+                description: 'Table custom footer',
+                props: ''
             }
         ],
         events: [