+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * **New component: Collapse**, [see docs](https://buefy.github.io/#/documentation/collapse)
 * **Panel deprecated**, use new Collapse component, [old docs here](https://buefy.github.io/#/documentation/panel)
 * Update Bulma to v0.5.3
+* #223 Add ``footer`` slot to table
 * #246 Remove ``key`` on detailed rows of table component (thanks @Tirke)
 * #301 Fix Bulma style override for radio and checkbox
 * #303 Fix dropdown close on custom item when is replaced or changed